test: 상품 재고 테스트 코드 개발

+++ module-core/src/main/kotlin/io/soboro/supreme/core/model/product/entity/Product.kt
@@ -28,9 +28,11 @@ class Product(
     @Column var price: Money,
 ) : BaseEntity() {
 
-    fun hasAvailableStock(itemId: Long, quantity: Int): Boolean {
-        val productItem = (this.productItems.find { item -> item.id == itemId }
-            ?: throw ProductNotFoundException())
+    fun hasAvailableStock(itemId: Long, quantity: Long): Boolean {
+        val productItem = (
+            this.productItems.find { item -> item.id == itemId }
+                ?: throw ProductNotFoundException()
+            )
 
         return productItem.hasEnough(quantity)
     }
